requirements.txt
streamlit
pandas
plotly
google-cloud-storage
